Finish up the conditionalization of the babysitter tests.

+++ test/nginx_system_test.sh
@@ -313,6 +313,7 @@ SERVER_NAME=nginx
 # thought as we don't currently have a multi-phase test flow in ngx_pagespeeed,
 # as far as I can tell, and we really have to update the config and restart
 # since the controller is a global setting.
+RUN_CONTROLLER_TEST=${RUN_CONTROLLER_TEST:-off}
 
 # run generic system tests
 SYSTEM_TEST_FILE="$MOD_PAGESPEED_DIR/src/pagespeed/system/system_test.sh"
@@ -607,28 +608,6 @@ function find_exactly_once {
   test $(grep -c "$1") -eq 1
 }
 
-function check_process_names() {
-  if ! $USE_VALGRIND; then
-    # There should be one babysitter and controller running.  Under valgrind
-    # process labels are confused, so skip the check then.
-
-    running=$(ps auxww | grep 'ngin[x]')
-    check_from "$running" find_exactly_once "nginx: pagespeed babysitter"
-    check_from "$running" find_exactly_once "nginx: pagespeed controller"
-  fi
-}
-
-check_process_names
-
-# Fire up some heavy load if ab is available to test a stressed reload.
-# TODO(oschaaf): make sure we wait for the new worker to get ready to accept
-# requests.
-fire_ab_load
-
-check wget $EXAMPLE_ROOT/styles/W.rewrite_css_images.css.pagespeed.cf.Hash.css \
-  -O /dev/null
-check_simple "$NGINX_EXECUTABLE" -s reload -c "$PAGESPEED_CONF"
-
 # Wait for the new worker process with the new configuration to get ready, or
 # else the sudden reset of the shared mem statistics/cache might catch upcoming
 # tests unaware.
@@ -638,21 +617,45 @@ function wait_for_new_worker() {
     sleep .1
   done
 }
-wait_for_new_worker
-check wget $EXAMPLE_ROOT/styles/W.rewrite_css_images.css.pagespeed.cf.Hash.css \
-  -O /dev/null
-if [ "$AB_PID" != "0" ]; then
-    echo "Kill ab (pid: $AB_PID)"
-    kill -s KILL $AB_PID &>/dev/null || true
+
+if [ "$RUN_CONTROLLER_TEST" = "on" ]; then
+  function check_process_names() {
+    if ! $USE_VALGRIND; then
+      # There should be one babysitter and controller running.  Under valgrind
+      # process labels are confused, so skip the check then.
+
+      running=$(ps auxww | grep 'ngin[x]')
+      check_from "$running" find_exactly_once "nginx: pagespeed babysitter"
+      check_from "$running" find_exactly_once "nginx: pagespeed controller"
+    fi
+  }
+
+  check_process_names
+
+  # Fire up some heavy load if ab is available to test a stressed reload.
+  # TODO(oschaaf): make sure we wait for the new worker to get ready to accept
+  # requests.
+  fire_ab_load
+
+  URL="$EXAMPLE_ROOT/styles/W.rewrite_css_images.css.pagespeed.cf.Hash.css"
+  check wget "$URL" -O /dev/null
+  check_simple "$NGINX_EXECUTABLE" -s reload -c "$PAGESPEED_CONF"
+
+  wait_for_new_worker
+  check wget "$URL" -O /dev/null
+  if [ "$AB_PID" != "0" ]; then
+      echo "Kill ab (pid: $AB_PID)"
+      kill -s KILL $AB_PID &>/dev/null || true
+  fi
+
+  # There should still be just one babysitter and controller running.
+  check_process_names
+
+  check grep "Writing a byte to a pipe to tell the old controller to exit." \
+    $ERROR_LOG
+  check grep "Root process is starting a new controller; shutting down." \
+    $ERROR_LOG
 fi
-
-# There should still be just one babysitter and controller running.
-check_process_names
-
-check grep "Writing a byte to a pipe to tell the old controller to exit." \
-  $ERROR_LOG
-check grep "Root process is starting a new controller; shutting down." \
-  $ERROR_LOG
 
 start_test "Shared memory checkpointing"
